+++ src/test/java/gg/loto/user/service/UserUpdateProfileServiceTest.java
@@ -1,9 +1,9 @@
 package gg.loto.user.service;
 
-import gg.loto.user.dto.UserResponse;
-import gg.loto.user.entity.User;
+import gg.loto.user.web.dto.UserResponse;
+import gg.loto.user.domain.User;
 import gg.loto.user.repository.UserRepository;
-import gg.loto.user.dto.UserUpdateRequest;
+import gg.loto.user.web.dto.UserUpdateRequest;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
